Refactor: Subtle change is added
