+++ backend/api/api/router.go
@@ -1,13 +1,19 @@
-package routers
+package api
 
 import (
+	"database/sql"
 	"net/http"
 
 	"github.com/gorilla/mux"
 	"github.com/sns/backend/controllers"
+	"github.com/sns/backend/services"
 )
 
-func NewRouter(aCon *controllers.ArticleController, cCon *controllers.CommentController) *mux.Router {
+func NewRouter(db *sql.DB) *mux.Router {
+	ser := services.NewMyAppService(db)
+	aCon := controllers.NewArticleController(ser)
+	cCon := controllers.NewCommentController(ser)
+
 	r := mux.NewRouter()
 
 	r.HandleFunc("/article", aCon.PostArticleHandler).Methods(http.MethodPost)
