+++ backend/api/main.go
@@ -7,9 +7,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/sns/backend/controllers"
-	"github.com/sns/backend/routers"
-	"github.com/sns/backend/services"
+	"github.com/sns/backend/api"
 
 	_ "github.com/go-sql-driver/mysql"
 )
@@ -29,11 +27,7 @@ func main() {
 		log.Println("fail to connect DB")
 	}
 
-	ser := services.NewMyAppService(db)
-	aCon := controllers.NewArticleController(ser)
-	cCon := controllers.NewCommentController(ser)
-
-	r := routers.NewRouter(aCon, cCon)
+	r := api.NewRouter(db)
 
 	log.Println("server start at port 8080")
 	log.Fatal(http.ListenAndServe(":8080", r))
